Fix tsc errors: add BaseExportSvgDialog, ErrorBanner alias, fix FileSaver and feats type
- Add BaseExportSvgDialog to core/ui, use it in circular-view and dotplot ExportSvgDialog
- Export ErrorMessage as ErrorBanner alias from core/ui
- Fix FileSaver: blob.name only exists on File, not Blob
- Cast feats to SimpleFeature[] in ChordVariantDisplay stateModelFactory
